Refactor: 동일한 QueryClient 인스턴스 사용으로 캐시와 상태의 일관성 유지시킴 - 서버 컴포넌트에서도 사용 가능하도록 getQueryClient 분리 - FeedPage에 new QueryClient 대신 getQueryClient 사용
