+++ src/app/api/posts/[postId]/comments/[commentId]/route.ts
@@ -1,4 +1,4 @@
-import { db } from "@/config/firebase";
+import { db } from "@/config/firebase/firebase";
 import { doc, updateDoc, deleteDoc, serverTimestamp } from "firebase/firestore";
 import { NextRequest, NextResponse } from "next/server";
