+++ src/app/api/posts/[postId]/route.ts
@@ -1,4 +1,4 @@
-import { db, storage } from "@/config/firebase";
+import { db, storage } from "@/config/firebase/firebase";
 import {
   deleteDoc,
   doc,
