+++ src/config/tanstack-query/get-query-client.ts
@@ -1,0 +1,26 @@
+import { QueryClient } from "@tanstack/react-query";
+
+function makeQueryClient() {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: 60 * 1000,
+      },
+    },
+  });
+}
+
+let browserQueryClient: QueryClient | undefined = undefined;
+
+export default function getQueryClient() {
+  if (typeof window === "undefined") {
+    // Server: 항상 새로운 queryClient 생성
+    return makeQueryClient();
+  } else {
+    // Browser: 다시 만들지 않고 기존에 이미 client 존재시 해당 client 제공
+    if (!browserQueryClient) {
+      browserQueryClient = makeQueryClient();
+    }
+    return browserQueryClient;
+  }
+}
