+++ src/lib/auth/api.ts
@@ -1,4 +1,4 @@
-import { auth, db } from "@/config/firebase";
+import { auth, db } from "@/config/firebase/firebase";
 import {
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
