+++ src/providers/QueryProvider.tsx
@@ -1,33 +1,8 @@
 "use client";
 
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import getQueryClient from "@/config/tanstack-query/get-query-client";
+import { QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-
-function makeQueryClient() {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: 60 * 1000,
-      },
-    },
-  });
-}
-
-let browserQueryClient: QueryClient | undefined = undefined;
-
-function getQueryClient() {
-  if (typeof window === "undefined") {
-    // Server: 항상 새로운 queryClient 생성
-    return makeQueryClient();
-  } else {
-    // Browser: 다시 만들지 않고 기존에 이미 client 존재시 해당 client 제공
-    if (!browserQueryClient) {
-      browserQueryClient = makeQueryClient();
-    }
-    return browserQueryClient;
-  }
-}
-// const queryClient = new QueryClient();
 
 export default function QueryProvider({
   children,
